requirements.txt
Flask-Session
pillow
